Change 'com' to 'io' in namespace
