+++ pom.xml
@@ -2,7 +2,7 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>com.github.cjengineer18</groupId>
+	<groupId>io.github.cjengineer18</groupId>
 	<artifactId>desktop-window-template</artifactId>
 	<description>JFrame for dummies. Creates desktop windows for Java.</description>
 	<version>1.3</version>
