+++ src/main/java/io/github/cjengineer18/desktopwindowtemplate/exception/UnknownAsyncProcessException.java
@@ -1,0 +1,46 @@
+/* 
+ * Copyright (c) 2018-2022 Cristian José Jiménez Diazgranados
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package io.github.cjengineer18.desktopwindowtemplate.exception;
+
+/**
+ * This exception is thrown when an {@link Exception} is thrown during an async
+ * process. However, the cause of the exception is unknown.
+ * 
+ * @see AsyncProcessException
+ * 
+ * @author Cristian Jimenez Dzg
+ *
+ */
+public class UnknownAsyncProcessException extends Exception {
+
+	private static final long serialVersionUID = -4586763719006993165L;
+
+	public UnknownAsyncProcessException() {
+		super("An unknown error was thrown during the process");
+	}
+
+	@Override
+	public String getMessage() {
+		return "An unknown error was thrown during the process";
+	}
+
+}
