+++ src/main/resources/io/github/cjengineer18/desktopwindowtemplate/resources/bundle/panels_es.properties
